+++ examples/client/src/main.rs
@@ -11,6 +11,7 @@ use netstack::{
         ConnectionToken
     },
     packets::OutgoingPacket,
+    monitoring::EmptyClientMonitor,
 };
 use std::net::SocketAddr;
 use std::time::Duration;
@@ -25,7 +26,7 @@ pub struct ConnectionInfo {
 }
 
 fn get_connection_info() -> ConnectionInfo {
-    let response = ureq::get("http://127.0.0.1:8000/").call();
+    let response = ureq::get("http://127.0.0.1:8000/token").call();
 
     if !response.ok() {
         panic!("could not get a secret from the remote server");
@@ -49,7 +50,9 @@ fn main() {
         heartbeat: 60,
     };
 
-    let mut client = Client::new(config, Box::new(transport));
+    let monitor = EmptyClientMonitor::new();
+
+    let mut client = Client::new(config, Box::new(transport), Box::new(monitor));
 
     let connection_info = get_connection_info();